fix: 401 - anon user for CreateSetNode requests
there are some trouble with the cookie/php session timeout.
sometimes after the cli runs for a long time further calls with the
same cookie will be rejected as "401 anonymous user".
Now we recreate the client for each feed to reduce this time.

currently verified if this was the real cause....